+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_params.py
@@ -160,7 +160,6 @@ def load_arguments(self, _):
         c.argument('vm_name', existing_vm_name)
         c.argument('size', completer=get_vm_size_completion_list)
         c.argument('name', arg_type=name_arg_type)
-        c.argument('computer_name')
         c.argument('zone', zone_type, min_api='2017-03-30')
         c.argument('caching', help='Disk caching policy', arg_type=get_enum_type(CachingTypes))
         c.argument('nsg', help='The name to use when creating a new Network Security Group (default) or referencing an existing one. Can also reference an existing NSG by ID or specify "" for none.', arg_group='Network')
@@ -460,8 +459,8 @@ def load_arguments(self, _):
             c.argument('generate_ssh_keys', action='store_true', help='Generate SSH public and private key files if missing. The keys will be stored in the ~/.ssh directory')
             c.argument('admin_username', help='Username for the VM.', default=get_default_admin_username())
             c.argument('admin_password', help="Password for the VM if authentication type is 'Password'.")
-            c.argument('ssh_key_value', help='SSH public key or public key file path.', completer=FilesCompleter(), type=file_type)
-            c.argument('ssh_dest_key_path', help='Destination file path on the VM for the SSH key.')
+            c.argument('ssh_key_value', completer=FilesCompleter(), type=file_type, nargs='*')
+            c.argument('ssh_dest_key_path', help='Destination file path on the VM for the SSH key. If the file already exists, the specified key(s) are appended to the file.')
             c.argument('authentication_type', help='Type of authentication to use with the VM. Defaults to password for Windows and SSH public key for Linux. "all" enables both ssh and password authentication. ', arg_type=get_enum_type(['ssh', 'password', 'all']))
 
         with self.argument_context(scope, arg_group='Storage') as c: